+++ benches/benchmark.rs
@@ -64,6 +64,7 @@ mod from_entries;
 mod from_pairs;
 mod geometric_mean;
 mod group_by;
+mod harmonic_mean;
 mod has_key;
 mod index_of;
 mod interleave;
@@ -241,6 +242,7 @@ fn all_benches(c: &mut Criterion) {
     from_pairs::benchmark_from_pairs(c);
     geometric_mean::benchmark_geometric_mean(c);
     group_by::benchmark_group_by(c);
+    harmonic_mean::benchmark_harmonic_mean(c);
     has_key::benchmark_has_key(c);
     index_of::benchmark_index_of(c);
     interleave::benchmark_interleave(c);